+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.5.0-rc.15",
-  genVersion: "2.845.1",
-  userAgent: "speakeasy-sdk/typescript 4.5.0-rc.15 2.845.1 2.0 @orq-ai/node",
+  sdkVersion: "4.5.0-rc.16",
+  genVersion: "2.845.5",
+  userAgent: "speakeasy-sdk/typescript 4.5.0-rc.16 2.845.5 2.0 @orq-ai/node",
 } as const;