+++ packages/orq-rc/src/models/operations/createidentity.ts
@@ -139,7 +139,7 @@ export const CreateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-27T09:23:51.039Z",
+    "2026-02-27T11:09:03.160Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {